Kill this new and insanely annoying warning
Nikita, if this is not the right way to go about it, then please amend it

+++ Zend/zend_API.h
@@ -776,7 +776,7 @@ static zend_always_inline zval *zend_try_array_init(zval *zv) {
 
 static zend_always_inline zval *zend_try_array_init_size(zval *zv, size_t size) {
 	zval tmp;
-	ZVAL_ARR(&tmp, zend_new_array(size));
+	ZVAL_ARR(&tmp, zend_new_array((uint32_t)size));
 	if (UNEXPECTED(zend_try_assign(zv, &tmp) == FAILURE)) {
 		return NULL;
 	}
